add BluetoothScanStreamHandler
